generation/requirements.txt
aiohttp>=3.8.0
python-dotenv>=1.0.0
tqdm>=4.64.0